Remove model and use controller as entry, rename MainView
