+++ src/ui/Controller.java
@@ -6,20 +6,36 @@ import floor.Floor;
 import floor.Floor.directionLampState;
 import floor.FloorSubsystem;
 import floor.SimulationVars;
+import elevator.ElevatorManager;
+import floor.FloorManager;
 
 public class Controller {
-	private MainView mainView;
+	private View view;
+	private Thread elevatorManager;
+	private Thread floorManager;
 
 	public Controller() {
-		mainView = new MainView();
-		mainView.display();
+		view = new View();
+		elevatorManager = new Thread(new ElevatorManager(SimulationVars.numberOfElevators, this));
+		floorManager = new Thread(new FloorManager(this));
 	}
 
+    public void run() {
+		elevatorManager.start();
+		floorManager.start();
+		view.display();
+    }
+
 	public void updateElevator(Elevator elevator) {
-        mainView.getElevatorsView().render(elevator);
+        view.getElevatorsView().render(elevator);
 	}
 
 	public void updateFloor(FloorSubsystem floor) {
-        //mainView.renderFloor(floor);
+        //view.renderFloor(floor);
+	}
+
+	public static void main(String[] args) {
+		Controller c = new Controller();
+		c.run();
 	}
 }
